feat(left-sidebar): merge map constants for icons
